// airline-system-always-online-web-app/src/app/models/list.ts
export class List {

    id: string;

    price: string;

    CFlightTotalDuration: string;

    CfirstFlightAirline: string;

    CfirstFlightAirlineLogo: string;

    CfirstFlightArrivalAirport: string;

    CfirstFlightArrivalDate: string;

    CfirstFlightArrivalDuration: string;

    CfirstFlightDepartureAirport: string;

    CfirstFlightDepartureDate: string;

    CsecondFlightAirline: string;

    CsecondFlightAirlineLogo: string;

    CsecondFlightArrivalAirport: string;

    CsecondFlightArrivalDate:string;

    CsecondFlightArrivalDuration:string;

    CsecondFlightDepartureAirport: string;

    CsecondFlightDepartureDate: string;

    Cstop_time: string;

    OFlightTotalDuration: string;

    OfirstFlightAirline: string;

    OfirstFlightAirlineLogo: string;

    OfirstFlightArrivalAirport: string;

    OfirstFlightArrivalDate: string;

    OfirstFlightArrivalDuration: string;

    OfirstFlightDepartureAirport: string;

    OfirstFlightDepartureDate: string;

    OsecondFlightAirline: string;

    OsecondFlightAirlineLogo: string;

    OsecondFlightArrivalAirport: string;

    OsecondFlightArrivalDate:string;

    OsecondFlightArrivalDuration:string;

    OsecondFlightDepartureAirport: string;

    OsecondFlightDepartureDate: string;
    
    Ostop_time: string;
}