// airline-system-always-online-web-app/src/app/models/booking.ts
export class Booking {

    id: number;

    bookedDate: any;

    travelerId: number;

    originLocationCode: string;

    destinationLocationCode: string;

    departureDate: string;

    returnDate: string;

    adults: number;

    childs: number;

    totalPrice: number;

    firstFlightAirline: string;
}